+++ sources/samples/generate-ssl-key.sh
@@ -1,2 +1,5 @@
-openssl req -x509 -newkey rsa:2048 -nodes -sha256 -subj '//CN=localhost' \
-  -keyout localhost-privkey.pem -out localhost-cert.pem
+openssl req -x509 -newkey rsa:2048 -nodes -sha256 -subj '//CN=a.local.org' \
+  -keyout a.local.org-privkey.pem -out a.local.org-cert.pem
+
+openssl req -x509 -newkey rsa:2048 -nodes -sha256 -subj '//CN=b.local.org' \
+  -keyout b.local.org-privkey.pem -out b.local.org-cert.pem